cav-dados/projeto-finance/pyproject.toml
[tool.poetry]
name = "projeto-finance"
version = "0.1.0"
description = ""
authors = ["Fabricio Zillig <fabricio.zillig@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.10.0"
kafka-python-ng = "^2.2.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.1"
numpy = "^2.1.1"
torch = "^2.4.1"
transformers = "^4.45.1"
yfinance = "^0.2.44"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
